fix VC8 warning
git-svn-id: http://vtp.googlecode.com/svn/trunk@7042 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainApps/VTBuilder/vtImage.cpp
@@ -833,7 +833,7 @@ void vtImage::GetRGB(int x, int y, RGBi &rgb, double dRes)
 	else if (bm.m_bOverlay)
 	{
 		// support for out-of-memory image here
-		RGBi *data = m_linebuf.GetScanlineFromBuffer(y, closest_bitmap);
+		RGBi *data = m_linebuf.GetScanlineFromBuffer(y, (int)closest_bitmap);
 		rgb = data[x];
 	}
 }
